feat(WindowSystem): add window system management with creation, VSync and event management
